requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
flask-login==0.6.3